requirements.txt
google_api_python_client==2.191.0
google_auth_oauthlib==1.3.0
protobuf==7.34.0
python-dotenv==1.2.2